+++ Backend/src/main/java/com/project/backend/api/controller/RequestName.java
@@ -1,0 +1,5 @@
+package com.project.backend.api.controller;
+
+public enum RequestName {
+	DELETE;
+}
