+++ Backend/src/main/java/com/project/backend/api/model/request/LoginRequestModel.java
@@ -1,4 +1,4 @@
-package com.project.backend.ui.model.request;
+package com.project.backend.api.model.request;
 
 public class LoginRequestModel {
