+++ Backend/src/main/java/com/project/backend/api/model/response/ErrorMessage.java
@@ -1,4 +1,4 @@
-package com.project.backend.ui.model.response;
+package com.project.backend.api.model.response;
 
 import java.util.Date;
