+++ Backend/src/main/java/com/project/backend/api/model/response/RequestStatus.java
@@ -1,0 +1,5 @@
+package com.project.backend.api.model.response;
+
+public enum RequestStatus {
+	ERROR, SUCCESS;
+}
