--- Backend/src/main/java/com/project/backend/ui/controller/RequestName.java
@@ -1,5 +1,0 @@
-package com.project.backend.ui.controller;
-
-public enum RequestName {
-	DELETE;
-}
