--- Backend/src/main/java/com/project/backend/ui/model/response/RequestStatus.java
@@ -1,5 +1,0 @@
-package com.project.backend.ui.model.response;
-
-public enum RequestStatus {
-	ERROR, SUCCESS;
-}
